Fixed CreativeModeInventoryScreen issue
When the compost bag was in the first hotbar slot, the mouse was held down, and the screen was specifically the CreativeModeInventoryScreen, the carried slot would immediately be overwritten with air once the tick method returned true. We now check if it's that screen, and since the client always takes precedence for this specific screen, we instead call the client side menu clicked method. Nothing is sent to the server in this case until it is updated from the creative screen, which resolves this issue.