Fixed CreativeModeInventoryScreen issue Pt.2
So, slots are weird. The ItemPicker slots give you the proper index but when in the inventory tab it is always zero. getContainerSlot is supposed to return the index relative to the part of the container, so in the ItemPicker menu the 3rd slot would return 3 and in the hotbar the 3rd slot would also return 3, while the overall slot index would be 3 and then 47 respectively. However, for the inventory tab, the SlotWrapper takes the index and puts it in the slot field, which causes getContainerSlot to return the real index while index is 0. :thonk: