Added FABRIC QOL methods and syncing config
Syncing works on server join, server disconnect, and server reload. On server join, the client's values are overridden by the server. On server disconnect, the config is reloaded on the client side and the values are set. On server reload, the PlayerListMixin#reloadResources method broadcasts the new server values to every connected client.